+++ models/instance_settings.go
@@ -7,7 +7,7 @@ type InstanceSettings struct {
 	NavigationMode    NavigationMode   `bun:"navigation_mode,type:int"`
 	NavigationMethod  NavigationMethod `bun:"navigation_method,type:int"`
 	MaxNextLocations  int              `bun:"max_next_locations,type:int,default:3"`
-	CompletionMethod  CompletionMethod `bun:"completion_method,type:int"`
+	MustCheckOut      bool             `bun:"must_check_out,type:bool"`
 	ShowTeamCount     bool             `bun:"show_team_count,type:bool"`
 	EnablePoints      bool             `bun:"enable_points,type:bool"`
 	EnableBonusPoints bool             `bun:"enable_bonus_points,type:bool"`